Create git-commands-comparison
added a new file to the repo. This file has the unedited text from the content section of the assignment. I added this file directly on the GitHub repo using the site. All of the changes from now will be made from a local copy on my laptop via pull request.