Update git-commands-comparison
I decided to name this article "How to use git commands to view and update tracking branches" because that seemed to be what all three of the git commands had in common. I then used an H1 to display the title of the article and updated the article name and description in the metadata. I then turned header with the title for the section explaining the differences to an H2. I also separated the command examples and explanations and placed them into their own H2. I kept that before the explanation because they would need to know the commands before talking about the differences. I separated them since they were doing two different things, 1) explaining what the command was and 2) explaining the difference between commands. I then separated the text explaining the differences by command to make it easier to digest for the reader and clear when a new command was being explained.