updated text and added command to list
I removed "under the hood" because that's an expression that may not translate well to readers from other cultures. I also tried to simplify things by choosing smaller words. I also added contractions to make the sentences shorter. Defined a merge conflict. Added git merge command to the list of commands and defined it. I added the command since it was mentioned in the text but left out of the list of the commands, so the command would come as a surprise to the reader since it was excluded from the list unlike the rest of the commands
.