added an introduction
added information about the purpose of the article  and why the reader will need these commands. I think it's important to have this intro so that the reader will know early on if this is an article that will help them accomplish what it is they are trying to accomplish. So if the reader is someone who wants to know how they can get their changes into the main repo, they'll know from the intro that this is the right article to answer their question. I also defined tracking branches. Changed the title again since this isn't only about tracking branches; local branches and master branches are included as well.