updated content for clarity and cleanliness
Updated text in "What is the difference between push, pull, and fetch" section for clarity. Changed "master" to "main" throughout document. GitHub has added the command main to replace master to be less offensive to users.  Capitalized all instances of "Git" when it's used as a noun and not a command. Added punctuation after Git command definitions